--- apps/client/src/pages/memo/index.ts
@@ -1,1 +1,0 @@
-export { default as MemoPage } from './memo-page';
