+++ apps/client/src/pages/memo/memo-page.css.ts
@@ -1,10 +1,8 @@
 import { style } from '@vanilla-extract/css';
 
-import { themeVars } from '@cds/ui';
-
-export const container = style({
-  ...themeVars.fontStyles.display_sb_36,
-  backgroundImage: themeVars.color.gradient02,
-  color: themeVars.color.grey700,
-  width: '30rem',
+export const pageContainer = style({
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
+  height: '100vh',
 });
