+++ apps/client/src/pages/memo/memo-page.tsx
@@ -1,45 +1,7 @@
-import { PATH } from '@router/path';
-import { useNavigate, useSearchParams } from 'react-router';
+import * as styles from './memo-page.css';
 
-import EmptyView from '@shared/components/empty-view/empty-view';
-import MemoListView from '@shared/components/memo-list-view/memo-list-view';
-
-import { useGetAllMemo, useGetMemoTotalCount } from './apis/queries';
-
-import emptyImage from '/empty.svg';
-
-const AllMemoPage = () => {
-  const navigate = useNavigate();
-  const [searchParams] = useSearchParams();
-  const tagParam = searchParams.get('tag');
-  const tagId = tagParam !== null ? [Number(tagParam)] : undefined;
-
-  const {
-    data: filteredMemos,
-    hasNextPage,
-    isFetchingNextPage,
-    fetchNextPage,
-  } = useGetAllMemo(tagId);
-  const { data: totalCount } = useGetMemoTotalCount(tagId);
-
-  return totalCount === 0 ? (
-    <EmptyView
-      imgSrc={emptyImage}
-      title="작성된 메모가 없습니다."
-      description="새 메모 창에 들어가서 새로운 메모를 생성해보세요."
-      buttonText="메모 작성하러 가기"
-      onButtonClick={() => navigate(PATH.NEW_MEMO)}
-    />
-  ) : (
-    <MemoListView
-      title="전체 메모"
-      initialMemos={filteredMemos}
-      hasNextPage={hasNextPage}
-      isFetchingNextPage={isFetchingNextPage}
-      fetchNextPage={fetchNextPage}
-      totalCount={totalCount ?? 0}
-    />
-  );
+const NewMemoPage = () => {
+  return <div className={styles.pageContainer}></div>;
 };
 
-export default AllMemoPage;
+export default NewMemoPage;
