+++ apps/client/src/pages/memos/memos-page.css.ts
@@ -1,0 +1,10 @@
+import { style } from '@vanilla-extract/css';
+
+import { themeVars } from '@cds/ui';
+
+export const container = style({
+  ...themeVars.fontStyles.display_sb_36,
+  backgroundImage: themeVars.color.gradient02,
+  color: themeVars.color.grey700,
+  width: '30rem',
+});
