--- apps/client/src/pages/new-memo/index.ts
@@ -1,1 +1,0 @@
-export { default as NewMemoPage } from './new-memo-page';
