--- apps/client/src/pages/new-memo/new-memo-page.css.ts
@@ -1,8 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const pageContainer = style({
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  height: '100vh',
-});
