--- apps/client/src/pages/new-memo/new-memo-page.tsx
@@ -1,7 +1,0 @@
-import * as styles from './new-memo-page.css';
-
-const NewMemoPage = () => {
-  return <div className={styles.pageContainer}></div>;
-};
-
-export default NewMemoPage;
